// interfaces/SearchListing.d.ts
export interface SearchListing {
    type: string | sting[];
    location: string | string[],
    pickUp: string | string[],
    dropOff: string | string[],
    price_per_day: string | string[]
    minPrice: string | string[],
    maxPrice: string | string[],
    startDate: string | string[],
    endDate: string | string[]
}